See desc
* better multithreaded chunk generation (reuse threads, lock vertical
rows of the map only)
* 'y' on escape now exists back to server list
* vbos are used instead of display lists on opengl atleast 2.0
* added missing mutex locks that lead to crashes
* network.c: copy correct gamestate instead of whole union
* added check for texture size limitation
* greedy meshing "broken" for now, so don't use

+++ src/map.h
@@ -17,7 +17,7 @@
     along with BetterSpades.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-extern unsigned long long* map_colors;
+extern unsigned int* map_colors;
 extern int map_size_x;
 extern int map_size_y;
 extern int map_size_z;
@@ -44,7 +44,7 @@ unsigned long long map_get(int x, int y, int z);
 unsigned long long map_get_unblocked(int x, int y, int z);
 void map_set(int x, int y, int z, unsigned long long color);
 int map_cube_line(int x1, int y1, int z1, int x2, int y2, int z2, struct Point* cube_array);
-void map_vxl_setgeom(int x, int y, int z, unsigned int t, unsigned long long* map);
-void map_vxl_setcolor(int x, int y, int z, unsigned int t, unsigned long long* map);
-void map_vxl_load(unsigned char* v, unsigned long long* map);
+void map_vxl_setgeom(int x, int y, int z, unsigned int t, unsigned int* map);
+void map_vxl_setcolor(int x, int y, int z, unsigned int t, unsigned int* map);
+void map_vxl_load(unsigned char* v, unsigned int* map);
 void map_collapsing_render(float dt);
